requirements.txt

Django
django-ckeditor
django-cors-headers
django-crispy-forms
django-filter
django-js-asset
django-storages
djangorestframework
djangorestframework-simplejwt
docutils
filelock
Flask
Flask-Session
gunicorn
idna
install
isort
itsdangerous
Jinja2
jmespath
lazy-object-proxy
MarkupSafe
mccabe
Pillow
psycopg2
pycodestyle
PyJWT
pylint
python-dateutil
python-dotenv
pytz
requests
s3transfer
six

sqlparse
termcolor
toml
urllib3

Werkzeug
whitenoise
wrapt